Lift execute up to Py4JBackend
